add submit_tx command
